implemented server starting as external process with function
