+++ .vscode/launch.json
@@ -1,32 +1,32 @@
 {
-    // Use IntelliSense to learn about possible attributes.
-    // Hover to view descriptions of existing attributes.
-    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-    "version": "0.2.0",
-    "configurations": [
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Launch Program",
-            "program": "${workspaceFolder}/lib/server.js"
-        },
-        {
-            "type": "node",
-            "request": "launch",
-            "name": "Debug Unit Tests",
-            "port": 9230,
-            "runtimeArgs": [
-                "--inspect-brk=9230",
-                "${workspaceRoot}/node_modules/jest/bin/jest.js",
-                "--runInBand",
-                "--watch"
-            ],
-            "runtimeExecutable": null,
-            "console": "externalTerminal",
-            "internalConsoleOptions": "neverOpen",
-            "env": {
-                "BABEL_ENV": "testing"
-            }
-        }
-    ]
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Launch Program",
+      "program": "${workspaceFolder}/lib/server-launcher.js"
+    },
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "Debug Unit Tests",
+      "port": 9230,
+      "runtimeArgs": [
+        "--inspect-brk=9230",
+        "${workspaceRoot}/node_modules/jest/bin/jest.js",
+        "--runInBand",
+        "--watch"
+      ],
+      "runtimeExecutable": null,
+      "console": "externalTerminal",
+      "internalConsoleOptions": "neverOpen",
+      "env": {
+        "BABEL_ENV": "testing"
+      }
+    }
+  ]
 }
