+++ .vscode/settings.json
@@ -1,5 +1,12 @@
 {
     "cSpell.words": [
+        "Dependencies",
+        "badge",
+        "coverage",
+        "jest",
+        "mock",
+        "prettier",
+        "reflect",
         "splitted"
     ]
 }
