+++ package.json
@@ -1,24 +1,33 @@
 {
   "name": "ov-language-server",
-  "version": "0.0.5",
-  "license": "MIT",
+  "version": "0.0.5-typings4",
+  "description": "language-server implementation for openVALIDATION",
   "repository": {
     "type": "git",
     "url": "https://github.com/openvalidation/ov-language-server.git"
   },
-  "description": "language-server implementation for openVALIDATION",
+  "license": "MIT",
+  "scripts": {
+    "build": "tsc && npm run copy:exe",
+    "clean": "rimraf lib",
+    "copy:exe": "cpx ./src/rest-interface/ov-rest.exe ./lib/rest-interface/",
+    "prepare": "npm run clean && npm run build",
+    "start": "npm run build && node lib/server-launcher.js",
+    "start:ext": "node lib/server-launcher.js --external && npm run build",
+    "start:watch": "nodemon",
+    "test": "jest && make-coverage-badge",
+    "test:watch": "jest --watch"
+  },
   "dependencies": {
     "axios": "^0.19.0",
     "class-transformer": "^0.2.3",
-    "cpx": "^1.5.0",
-    "express": "^4.15.2",
-    "js-yaml": "^3.13.1",
-    "lodash": "^4.17.13",
-    "make-coverage-badge": "^1.1.1",
     "reflect-metadata": "^0.1.13",
+    "typescript": "^3.6.4",
     "typescript-string-operations": "^1.3.1",
     "vscode-languageserver": "^5.2.1",
     "vscode-languageserver-types": "^3.14.0",
+    "js-yaml": "^3.13.1",
+    "lodash": "^4.17.13",
     "vscode-uri": "^2.0.3",
     "vscode-ws-jsonrpc": "^0.1.1",
     "ws": "^7.1.2"
@@ -32,38 +41,16 @@
     "@types/normalize-url": "^4.2.0",
     "@types/ws": "^6.0.3",
     "axios-mock-adapter": "^1.17.0",
+    "cpx": "^1.5.0",
+    "express": "^4.15.2",
     "jest": "^24.9.0",
+    "make-coverage-badge": "^1.1.1",
     "nodemon": "^1.19.4",
     "prettier": "^1.18.2",
     "rimraf": "^3.0.0",
     "ts-jest": "^24.1.0",
     "tslint-config-prettier": "^1.18.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.6.4"
+    "tslint-plugin-prettier": "^2.0.1"
   },
-  "scripts": {
-    "prepare": "npm run clean && npm run build",
-    "clean": "rimraf lib",
-    "build": "tsc && npm run copy:exe",
-    "start": "npm run build && node lib/server.js",
-    "start:external": "npm run build && node lib/server.js --external",
-    "start:watch": "nodemon",
-    "test": "jest && make-coverage-badge",
-    "test:watch": "jest --watch",
-    "debug": "nodemon --debug lib/server.js",
-    "copy:exe": "cpx ./src/rest-interface/ov-rest.exe ./lib/rest-interface/"
-  },
-  "nodemonConfig": {
-    "ignore": [
-      "**/*.test.ts",
-      "**/*.spec.ts",
-      ".git",
-      "node_modules"
-    ],
-    "watch": [
-      "src"
-    ],
-    "exec": "npm start",
-    "ext": "ts"
-  }
+  "readmeFilename": "README.md"
 }
