+++ tsconfig.json
@@ -1,35 +1,26 @@
 {
-    "compilerOptions": {
-        "experimentalDecorators": true,
-        "target": "es2017",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "sourceMap": true,
-        "declarationMap": true,
-        "inlineSources": false,
-        "declaration": true,
-        "stripInternal": true,
-        "lib": [
-            "es2017",
-            "dom"
-        ],
-        "strict": true,
-        "noImplicitReturns": true,
-        "noUnusedLocals": true,
-        "rootDir": "src",
-        "outDir": "lib",
-        "baseUrl": "./",
-        "paths": {
-            "src/*": [
-                "src/*"
-            ],
-            "test/*": [
-                "test/*"
-            ]
-        },
-        "skipLibCheck": true
+  "compilerOptions": {
+    "experimentalDecorators": true,
+    "target": "es2017",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "sourceMap": true,
+    "declarationMap": true,
+    "inlineSources": false,
+    "declaration": true,
+    "stripInternal": true,
+    "lib": ["es2017", "dom"],
+    "strict": true,
+    "noImplicitReturns": true,
+    "noUnusedLocals": true,
+    "rootDir": "src",
+    "outDir": "lib",
+    "baseUrl": "./",
+    "paths": {
+      "src/*": ["src/*"],
+      "test/*": ["test/*"]
     },
-    "include": [
-        "src"
-    ]
+    "skipLibCheck": true
+  },
+  "include": ["src"]
 }
